+++ src/main/kotlin/net/ccbluex/liquidbounce/utils/aiming/features/anglesmooth/AngleSmoothMode.kt
@@ -22,20 +22,8 @@
 package net.ccbluex.liquidbounce.utils.aiming.features.anglesmooth
 
 import net.ccbluex.liquidbounce.config.types.Choice
-import net.ccbluex.liquidbounce.utils.aiming.data.Rotation
-import net.minecraft.entity.Entity
-import net.minecraft.util.math.Vec3d
 
 /**
- * A smoother is being used to limit the angle change between two rotations.
+ * An [AngleSmooth]'er, but as choice
  */
-abstract class AngleSmoothMode(name: String) : Choice(name) {
-    abstract fun limitAngleChange(
-        factorModifier: Float,
-        currentRotation: Rotation,
-        targetRotation: Rotation,
-        vec3d: Vec3d? = null,
-        entity: Entity? = null
-    ): Rotation
-    abstract fun howLongToReach(currentRotation: Rotation, targetRotation: Rotation): Int
-}
+abstract class AngleSmoothMode(name: String) : Choice(name), AngleSmooth
